Move blockchain service metrics package (#4959)
* Moved metrics package
