+++ beacon-chain/blockchain/BUILD.bazel
@@ -8,6 +8,7 @@ go_library(
         "info.go",
         "init_sync_process_block.go",
         "log.go",
+        "metrics.go",
         "process_attestation.go",
         "process_attestation_helpers.go",
         "process_block.go",
@@ -19,7 +20,6 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/blockchain",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
-        "//beacon-chain/blockchain/metrics:go_default_library",
         "//beacon-chain/cache:go_default_library",
         "//beacon-chain/cache/depositcache:go_default_library",
         "//beacon-chain/core/blocks:go_default_library",
@@ -49,6 +49,8 @@ go_library(
         "//shared/traceutil:go_default_library",
         "@com_github_emicklei_dot//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
+        "@com_github_prometheus_client_golang//prometheus:go_default_library",
+        "@com_github_prometheus_client_golang//prometheus/promauto:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
         "@com_github_prysmaticlabs_go_ssz//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
