--- beacon-chain/blockchain/metrics/BUILD.bazel
@@ -1,17 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["metrics.go"],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/blockchain/metrics",
-    visibility = ["//beacon-chain:__subpackages__"],
-    deps = [
-        "//beacon-chain/core/epoch/precompute:go_default_library",
-        "//beacon-chain/state:go_default_library",
-        "//shared/bytesutil:go_default_library",
-        "//shared/params:go_default_library",
-        "@com_github_prometheus_client_golang//prometheus:go_default_library",
-        "@com_github_prometheus_client_golang//prometheus/promauto:go_default_library",
-        "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
-    ],
-)
